Move reqs-handling code here and adjust callers
Move all the requirements-related code (but not the requirements
themselves, that is, not RTM.xlsx) from the PSM repository to here.
Pull requirements directly from RTM.xlsx (which still lives in the PSM
repository) to build dashboard input.

This is related to https://github.com/SolutionGuidance/psm/issues/799:
we're able to stop storing reqs CSV files in the PSM repository
because the code here can now extract requirements from RTM.xlsx.