Move reqs-handling code here and adjust callers
Move all the requirements-related code (but not the requirements
themselves, that is, not RTM.xlsx) from the PSM repository to here.
Pull requirements directly from RTM.xlsx (which still lives in the PSM
repository) to build dashboard input.

This is related to https://github.com/SolutionGuidance/psm/issues/799:
we're able to stop storing reqs CSV files in the PSM repository
because the code here can now extract requirements from RTM.xlsx.

WIP: This doesn't quite work yet, and I'm not sure why.  If you visit
burn-down.html, everything looks completed on the pie chart and
everything looks not-started on the feature progress chart.  Also, in
dashboard-input.json, every "status" field is either "Completed" or
"NotStarted" -- literally nothing is "InProgress".  But I don't know
if that's because of the changes in this commit, or because of changes
made to the source spreadsheet (perhaps related to issue #5).  It's
quite possible that the code still builds the dashboard in a totally
equivalent way to how it did before this commit, but that that way
doesn't work any more due to changes in the source data, such that if
we had run the old code to refresh the input JSON file we still would
have gotten a similarly broken burn-down chart.  Investigation needed.