Standardize on one json filename
With this change, the chart will expect to be built using the same
filename we recommend as output from the get-inputs script.  Before,
this was silently pulling from the sample data, leading to confusion
about which was correct.

One question: should we version sample-output.json?

I made this change in b68b4ec and then reverted it in 4699db1 because it
broken master for me until the fix in 3262e3f.  Now it should all work
correctly!