Add toggle for feature descriptions, closes #12
Adds a toggle for displaying feature descriptions in each row, and also
increasing the overall height of the chart so they are visible. Also
uses lighter colors so that the text has more contrast over the rows
without having the alter the font color, and changes the overlay
position property to "fixed" so that it isn't stuck at the top of the
frame when scrolling.