Tolerate extra whitespace in req ID lists
This fixes an Elisp-ism in the regexp used for matching req ID lists.
In Emacs Lisp, "\s-" matches the whitespace class, but in Python it's
just "\s".  The presence of the hyphen meant that a list like this:

  (psm-FR-3.1,   psm-FR-5.8, psm-FR-6.4)

wouldn't match, even though we want it to.  On the other hand, with
only a single space, it would match:

  (psm-FR-3.1, psm-FR-5.8, psm-FR-6.4)

But the match would be purely by concidence -- it was because the next
thing in the regexp was a "*" to mean "0 or more", so even though the
feature descriptions in the spreadsheet never have hyphens there, we
would match anyway by matching, ahem, zero hyphens.

Now it matches any amount of whitespace, which is what we want.

The above was a real example, by the way -- as of this writing, it's
psm-feature-049, row 218 of the "PSM Features - RTM map" spreadsheet.
We could have just compacted the extra whitespace in that entry, but
that doesn't seem right: if extra whitespace got into an entry once,
it's going to happen again somewhere, so the code should tolerate it.