Add Ongoing status, closing #6, #25
Adds the "Ongoing" status from the spreadsheet to the dashboard by
defaulting to any value specified in the "Status" column before
attempting to set the status for a feature. Updates sorting logic to
compare "InProgress" and "NotStarted" features against each other
in the event of projected future dates. Updates the feature data with
the most recent information and "Ongoing" status.