Make the 'show-reqs' script work smoothly again
The 'show-reqs' script worked, but it was looking in an obsolete place
for the requirements CSV file.  Now it looks in the same place that
'refresh-dashboard' looks.  In order to share that common location, we
abstract out the CSV generation to a new script 'reqs2csv' (and
therefore rename the old 'reqs2any' script to 'format-reqs' -- the
"any" in "reqs2any" wrongly implied that it handled conversion to CSV,
when in fact CSV is its input; now that we have "reqs2csv", the name
really needed to be fixed).