Handle feature descriptions that have no reqs list
Fix a bug whereby a feature description that contained no requirements
list would end up with a blank description, and thus show as a blank
row in the dashboard when "Show Feature Descriptions" is turned on.

One such feature is 'psm-feature-000', whose description is

  "Continue state+CMS information-gathering and iterative feedback
   (Research for, e.g., psm-FR-1.3, psm-FR-2.10, psm-FR-2.12, psm-FR-5.1)"

While it lists some reqs, and they are inside parentheses, they are
just given as examples -- it is not a reqs list in the formal sense,
and it properly does not match the 'reqs_list_re' regexp in
'gather_info'.  That made 'psm-feature-000' the only one, so far, to
not have a formal reqs list, but there could be others in the future.