Remove the Y-axis label ("Features") since it's no longer helpful
Because the chart is now taller with the feature descriptions,
which pushed the Y-axis label down below the initial view, making
it less useful, and because now that we have a header describing
what the bars are at the top of the chart. Better to use that
horizontal space for the chart itself, especially on narrower
screens.