Remove frontend dashboard files and update README.md
The frontend files for the dashboard now live in the gh-pages
branch of the main psm repo. They were moved there in this commit:
https://github.com/SolutionGuidance/psm/commit/eea0afb94e5a338408fb3f69d6f1a96cd1bfe0ee

Thus, they no longer need to be in this psm-dashboard repo. The
features-info.json file will be generated by scripts in this
repo and checked in to the gh-pages branch of the main psm repo.
See issue #33 in this repo for more details.