Improve check for reqs spreadsheet RTM.xlsx
Similarly to commit 02d62399c1e, this follows up to commit ca5a69cee21
to restore an error-check that was lost from 'refresh-dashboard' in
the latter commit.  Furthermore, this properly assigns responsibility
for different kinds of checks between 'gather-info' and 'psm_reqs.py':
the former knows about 'psm-dashboard-config.json' and is therefore
entitled to complain about missing values in the config file; the
latter is a generic module that is ignorant of its caller's universe
and that should therefore give errors reflective of its narrower view.