Assume xlsx2csv was fetched via requirements.txt
As @pjsier pointed out in [1], commit ca5a69cee21 added xlsx2csv to
requirements.txt, so there's no need for commit 02d62399c1e to check
for error when trying to import xlsx2csv.  However, we do want to keep
the comment explaining both alternate sources for that library and
alternate libraries entirely, because the .xlsx conversion has been a
particularly fragile part of this whole process and it's quite likely
that someone may eventually need to either use the latest upstream
version or a different package entirely.

[1] https://github.com/SolutionGuidance/psm-dashboard/pull/27#issuecomment-414733751